all: refactor txpool into it's own package in prep for 4844 (#26038)
