+++ core/txpool/journal.go
@@ -14,9 +14,10 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package core
+package txpool
 
 import (
+	"errors"
 	"io"
 	"os"
 
@@ -26,23 +27,36 @@ import (
 	"github.com/XinFinOrg/XDPoSChain/rlp"
 )
 
-// ordertxJournal is a rotating log of transactions with the aim of storing locally
+// errNoActiveJournal is returned if a transaction is attempted to be inserted
+// into the journal, but no such file is currently open.
+var errNoActiveJournal = errors.New("no active journal")
+
+// devNull is a WriteCloser that just discards anything written into it. Its
+// goal is to allow the transaction journal to write into a fake journal when
+// loading transactions on startup without printing warnings due to no file
+// being readt for write.
+type devNull struct{}
+
+func (*devNull) Write(p []byte) (n int, err error) { return len(p), nil }
+func (*devNull) Close() error                      { return nil }
+
+// journal is a rotating log of transactions with the aim of storing locally
 // created transactions to allow non-executed ones to survive node restarts.
-type ordertxJournal struct {
+type journal struct {
 	path   string         // Filesystem path to store the transactions at
 	writer io.WriteCloser // Output stream to write new transactions into
 }
 
-// newOrderTxJournal creates a new transaction journal to
-func newOrderTxJournal(path string) *ordertxJournal {
-	return &ordertxJournal{
+// newTxJournal creates a new transaction journal to
+func newTxJournal(path string) *journal {
+	return &journal{
 		path: path,
 	}
 }
 
 // load parses a transaction journal dump from disk, loading its contents into
 // the specified pool.
-func (journal *ordertxJournal) load(add func(*types.OrderTransaction) error) error {
+func (journal *journal) load(add func([]*types.Transaction) []error) error {
 	// Skip the parsing if the journal file doens't exist at all
 	if _, err := os.Stat(journal.path); os.IsNotExist(err) {
 		return nil
@@ -62,22 +76,38 @@ func (journal *ordertxJournal) load(add func(*types.OrderTransaction) error) err
 	stream := rlp.NewStream(input, 0)
 	total, dropped := 0, 0
 
-	var failure error
+	// Create a method to load a limited batch of transactions and bump the
+	// appropriate progress counters. Then use this method to load all the
+	// journalled transactions in small-ish batches.
+	loadBatch := func(txs types.Transactions) {
+		for _, err := range add(txs) {
+			if err != nil {
+				log.Debug("Failed to add journaled transaction", "err", err)
+				dropped++
+			}
+		}
+	}
+	var (
+		failure error
+		batch   types.Transactions
+	)
 	for {
 		// Parse the next transaction and terminate on error
-		tx := new(types.OrderTransaction)
+		tx := new(types.Transaction)
 		if err = stream.Decode(tx); err != nil {
 			if err != io.EOF {
 				failure = err
 			}
+			if batch.Len() > 0 {
+				loadBatch(batch)
+			}
 			break
 		}
-		// Import the transaction and bump the appropriate progress counters
+		// New transaction parsed, queue up for later, import if threnshold is reached
 		total++
-		if err = add(tx); err != nil {
-			log.Debug("Failed to add journaled transaction", "err", err)
-			dropped++
-			continue
+		if batch = append(batch, tx); batch.Len() > 1024 {
+			loadBatch(batch)
+			batch = batch[:0]
 		}
 	}
 	log.Info("Loaded local transaction journal", "transactions", total, "dropped", dropped)
@@ -86,7 +116,7 @@ func (journal *ordertxJournal) load(add func(*types.OrderTransaction) error) err
 }
 
 // insert adds the specified transaction to the local disk journal.
-func (journal *ordertxJournal) insert(tx *types.OrderTransaction) error {
+func (journal *journal) insert(tx *types.Transaction) error {
 	if journal.writer == nil {
 		return errNoActiveJournal
 	}
@@ -98,7 +128,7 @@ func (journal *ordertxJournal) insert(tx *types.OrderTransaction) error {
 
 // rotate regenerates the transaction journal based on the current contents of
 // the transaction pool.
-func (journal *ordertxJournal) rotate(all map[common.Address]types.OrderTransactions) error {
+func (journal *journal) rotate(all map[common.Address]types.Transactions) error {
 	// Close the current journal (if any is open)
 	if journal.writer != nil {
 		if err := journal.writer.Close(); err != nil {
@@ -138,7 +168,7 @@ func (journal *ordertxJournal) rotate(all map[common.Address]types.OrderTransact
 }
 
 // close flushes the transaction journal contents to disk and closes the file.
-func (journal *ordertxJournal) close() error {
+func (journal *journal) close() error {
 	var err error
 
 	if journal.writer != nil {
