+++ core/txpool/lending_pool.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package core
+package txpool
 
 import (
 	"errors"
@@ -24,11 +24,12 @@ import (
 	"sync"
 	"time"
 
-	"github.com/XinFinOrg/XDPoSChain/XDCx/tradingstate"
+	"github.com/XinFinOrg/XDPoSChain/XDCxlending/lendingstate"
 	"github.com/XinFinOrg/XDPoSChain/common"
 	"github.com/XinFinOrg/XDPoSChain/common/prque"
 	"github.com/XinFinOrg/XDPoSChain/consensus"
 	"github.com/XinFinOrg/XDPoSChain/consensus/XDPoS"
+	"github.com/XinFinOrg/XDPoSChain/core"
 	"github.com/XinFinOrg/XDPoSChain/core/state"
 	"github.com/XinFinOrg/XDPoSChain/core/types"
 	"github.com/XinFinOrg/XDPoSChain/event"
@@ -37,35 +38,26 @@ import (
 )
 
 var (
-	// ErrInvalidOrderFormat is returned if the order transaction contains an invalid field.
-	ErrInvalidOrderFormat      = errors.New("invalid order format")
-	ErrInvalidOrderContent     = errors.New("invalid order content")
-	ErrInvalidOrderSide        = errors.New("invalid order side")
-	ErrInvalidOrderType        = errors.New("invalid order type")
-	ErrInvalidOrderStatus      = errors.New("invalid order status")
-	ErrInvalidOrderUserAddress = errors.New("invalid order user address")
-	ErrInvalidOrderQuantity    = errors.New("invalid order quantity")
-	ErrInvalidOrderPrice       = errors.New("invalid order price")
-	ErrInvalidOrderHash        = errors.New("invalid order hash")
-	ErrInvalidCancelledOrder   = errors.New("invalid cancel orderid")
+	ErrInvalidLendingSide        = errors.New("invalid lending side")
+	ErrInvalidLendingType        = errors.New("invalid lending type")
+	ErrInvalidLendingStatus      = errors.New("invalid lending status")
+	ErrInvalidLendingUserAddress = errors.New("invalid lending user address")
+	ErrInvalidLendingQuantity    = errors.New("invalid lending quantity")
+	ErrInvalidLendingInterest    = errors.New("invalid lending interest")
+	ErrInvalidLendingRelayer     = errors.New("invalid lending relayer address")
+	ErrInvalidLendingHash        = errors.New("invalid lending hash")
+	ErrInvalidCancelledLending   = errors.New("invalid cancel lending id")
+	ErrInvalidLendingTradeID     = errors.New("invalid lending trade ID")
+	ErrInvalidLendingCollateral  = errors.New("invalid collateral")
 )
 
 var (
-	OrderTypeLimit    = "LO"
-	OrderTypeMarket   = "MO"
-	OrderStatusNew    = "NEW"
-	OrderStatusCancle = "CANCELLED"
-	OrderSideBid      = "BUY"
-	OrderSideAsk      = "SELL"
+	LendingTypeLimit  = "LO"
+	LendingTypeMarket = "MO"
 )
 
-var (
-	ErrPendingNonceTooLow = errors.New("pending nonce too low")
-	ErrPoolOverflow       = errors.New("exceed pool size")
-)
-
-// OrderPoolConfig are the configuration parameters of the order transaction pool.
-type OrderPoolConfig struct {
+// LendingPoolConfig are the configuration parameters of the order transaction pool.
+type LendingPoolConfig struct {
 	NoLocals  bool          // Whether local transaction handling should be disabled
 	Journal   string        // Journal of local transactions to survive node restarts
 	Rejournal time.Duration // Time interval to regenerate the local transaction journal
@@ -79,12 +71,12 @@ type OrderPoolConfig struct {
 }
 
 // blockChain_XDCx add order state
-type blockChainXDCx interface {
+type blockChainLending interface {
 	CurrentBlock() *types.Block
 	GetBlock(hash common.Hash, number uint64) *types.Block
-	OrderStateAt(block *types.Block) (*tradingstate.TradingStateDB, error)
+	LendingStateAt(block *types.Block) (*lendingstate.LendingStateDB, error)
 	StateAt(root common.Hash) (*state.StateDB, error)
-	SubscribeChainHeadEvent(ch chan<- ChainHeadEvent) event.Subscription
+	SubscribeChainHeadEvent(ch chan<- core.ChainHeadEvent) event.Subscription
 	Engine() consensus.Engine
 	// GetHeader returns the hash corresponding to their hash.
 	GetHeader(common.Hash, uint64) *types.Header
@@ -94,9 +86,9 @@ type blockChainXDCx interface {
 	Config() *params.ChainConfig
 }
 
-// DefaultOrderPoolConfig contains the default configurations for the transaction
+// DefaultLendingPoolConfig contains the default configurations for the transaction
 // pool.
-var DefaultOrderPoolConfig = OrderPoolConfig{
+var DefaultLendingPoolConfig = LendingPoolConfig{
 	Journal:   "",
 	Rejournal: time.Hour,
 
@@ -110,74 +102,74 @@ var DefaultOrderPoolConfig = OrderPoolConfig{
 
 // sanitize checks the provided user configurations and changes anything that's
 // unreasonable or unworkable.
-func (config *OrderPoolConfig) sanitize() OrderPoolConfig {
+func (config *LendingPoolConfig) sanitize() LendingPoolConfig {
 	conf := *config
 	if conf.Rejournal < time.Second {
-		log.Warn("Sanitizing invalid OrderPool journal time", "provided", conf.Rejournal, "updated", time.Second)
+		log.Warn("Sanitizing invalid LendingPool journal time", "provided", conf.Rejournal, "updated", time.Second)
 		conf.Rejournal = time.Second
 	}
 	return conf
 }
 
-// OrderPool contains all currently known transactions. Transactions
+// LendingPool contains all currently known transactions. Transactions
 // enter the pool when they are received from the network or submitted
 // locally. They exit the pool when they are included in the blockchain.
 //
 // The pool separates processable transactions (which can be applied to the
 // current state) and future transactions. Transactions move between those
 // two states over time as they are received and processed.
-type OrderPool struct {
-	config      OrderPoolConfig
+type LendingPool struct {
+	config      LendingPoolConfig
 	chainconfig *params.ChainConfig
-	chain       blockChainXDCx
+	chain       blockChainLending
 
 	txFeed       event.Feed
 	scope        event.SubscriptionScope
-	chainHeadCh  chan ChainHeadEvent
+	chainHeadCh  chan core.ChainHeadEvent
 	chainHeadSub event.Subscription
-	signer       types.OrderSigner
+	signer       types.LendingSigner
 	mu           sync.RWMutex
 
-	currentRootState  *state.StateDB
-	currentOrderState *tradingstate.TradingStateDB   // Current order state in the blockchain head
-	pendingState      *tradingstate.XDCXManagedState // Pending state tracking virtual nonces
+	currentRootState    *state.StateDB
+	currentLendingState *lendingstate.LendingStateDB      // Current order state in the blockchain head
+	pendingState        *lendingstate.LendingManagedState // Pending state tracking virtual nonces
 
-	locals  *orderAccountSet // Set of local transaction to exempt from eviction rules
-	journal *ordertxJournal  // Journal of local transaction to back up to disk
+	locals  *lendingAccountSet // Set of local transaction to exempt from eviction rules
+	journal *lendingtxJournal  // Journal of local transaction to back up to disk
 
-	pending   map[common.Address]*ordertxList         // All currently processable transactions
-	queue     map[common.Address]*ordertxList         // Queued but non-processable transactions
-	beats     map[common.Address]time.Time            // Last heartbeat from each known account
-	all       map[common.Hash]*types.OrderTransaction // All transactions to allow lookups
-	wg        sync.WaitGroup                          // for shutdown sync
+	pending   map[common.Address]*lendingtxList         // All currently processable transactions
+	queue     map[common.Address]*lendingtxList         // Queued but non-processable transactions
+	beats     map[common.Address]time.Time              // Last heartbeat from each known account
+	all       map[common.Hash]*types.LendingTransaction // All transactions to allow lookups
+	wg        sync.WaitGroup                            // for shutdown sync
 	homestead bool
 	IsSigner  func(address common.Address) bool
 }
 
-// NewOrderPool creates a new transaction pool to gather, sort and filter inbound
+// NewLendingPool creates a new transaction pool to gather, sort and filter inbound
 // transactions from the network.
-func NewOrderPool(chainconfig *params.ChainConfig, chain blockChainXDCx) *OrderPool {
+func NewLendingPool(chainconfig *params.ChainConfig, chain blockChainLending) *LendingPool {
 	// Sanitize the input to ensure no vulnerable gas prices are set
-	config := (&DefaultOrderPoolConfig).sanitize()
-	log.Debug("NewOrderPool start...", "current block", chain.CurrentBlock().Header().Number)
+	config := (&DefaultLendingPoolConfig).sanitize()
+	log.Debug("NewLendingPool start...", "current block", chain.CurrentBlock().Header().Number)
 	// Create the transaction pool with its initial settings
-	pool := &OrderPool{
+	pool := &LendingPool{
 		config:      config,
 		chainconfig: chainconfig,
 		chain:       chain,
-		signer:      types.OrderTxSigner{},
-		pending:     make(map[common.Address]*ordertxList),
-		queue:       make(map[common.Address]*ordertxList),
+		signer:      types.LendingTxSigner{},
+		pending:     make(map[common.Address]*lendingtxList),
+		queue:       make(map[common.Address]*lendingtxList),
 		beats:       make(map[common.Address]time.Time),
-		all:         make(map[common.Hash]*types.OrderTransaction),
-		chainHeadCh: make(chan ChainHeadEvent, chainHeadChanSize),
+		all:         make(map[common.Hash]*types.LendingTransaction),
+		chainHeadCh: make(chan core.ChainHeadEvent, chainHeadChanSize),
 	}
-	pool.locals = newOrderAccountSet(pool.signer)
+	pool.locals = newLendingAccountSet(pool.signer)
 	pool.reset(nil, chain.CurrentBlock())
 
 	// If local transactions and journaling is enabled, load from disk
 	if !config.NoLocals && config.Journal != "" {
-		pool.journal = newOrderTxJournal(config.Journal)
+		pool.journal = newLendingTxJournal(config.Journal)
 
 		if err := pool.journal.load(pool.AddLocal); err != nil {
 			log.Warn("Failed to load transaction journal", "err", err)
@@ -199,10 +191,11 @@ func NewOrderPool(chainconfig *params.ChainConfig, chain blockChainXDCx) *OrderP
 // loop is the transaction pool's main event loop, waiting for and reacting to
 // outside blockchain events as well as for various reporting and transaction
 // eviction events.
-func (pool *OrderPool) loop() {
+func (pool *LendingPool) loop() {
 	defer pool.wg.Done()
 
 	// Start the stats reporting and transaction eviction tickers
+	var prevPending, prevQueued int
 
 	report := time.NewTicker(statsReportInterval)
 	defer report.Stop()
@@ -226,7 +219,7 @@ func (pool *OrderPool) loop() {
 				if pool.chainconfig.IsHomestead(ev.Block.Number()) {
 					pool.homestead = true
 				}
-				log.Debug("OrderPool new chain header reset pool", "old", head.Header().Number, "new", ev.Block.Header().Number)
+				log.Debug("LendingPool new chain header reset pool", "old", head.Header().Number, "new", ev.Block.Header().Number)
 				pool.reset(head, ev.Block)
 				head = ev.Block
 
@@ -241,8 +234,10 @@ func (pool *OrderPool) loop() {
 			pool.mu.RLock()
 			pending, queued := pool.stats()
 			pool.mu.RUnlock()
-
-			log.Debug("Order pool status report", "executable", pending, "queued", queued)
+			if pending != prevPending || queued != prevQueued {
+				log.Debug("Lending pool status report", "executable", pending, "queued", queued)
+				prevPending, prevQueued = pending, queued
+			}
 
 			// Handle inactive account transaction eviction
 		case <-evict.C:
@@ -276,25 +271,25 @@ func (pool *OrderPool) loop() {
 
 // reset retrieves the current state of the blockchain and ensures the content
 // of the transaction pool is valid with regard to the chain state.
-func (pool *OrderPool) reset(oldHead, newblock *types.Block) {
+func (pool *LendingPool) reset(oldHead, newblock *types.Block) {
 	if !pool.chainconfig.IsTIPXDCXReceiver(pool.chain.CurrentBlock().Number()) || pool.chain.Config().XDPoS == nil || pool.chain.CurrentBlock().NumberU64() <= pool.chain.Config().XDPoS.Epoch {
 		return
 	}
 	// If we're reorging an old state, reinject all dropped transactions
-	var reinject types.OrderTransactions
+	var reinject types.LendingTransactions
 
 	// Initialize the internal state to the current head
 	if newblock == nil {
 		newblock = pool.chain.CurrentBlock()
 	}
 	newHead := newblock.Header()
-	orderstate, err := pool.chain.OrderStateAt(newblock)
+	lendingState, err := pool.chain.LendingStateAt(newblock)
 	if err != nil {
-		log.Error("Failed to reset OrderPool state", "err", err)
+		log.Error("Failed to reset LendingPool state", "err", err)
 		return
 	}
-	pool.currentOrderState = orderstate
-	pool.pendingState = tradingstate.ManageState(orderstate)
+	pool.currentLendingState = lendingState
+	pool.pendingState = lendingstate.ManageState(lendingState)
 
 	state, err := pool.chain.StateAt(newHead.Root)
 	if err != nil {
@@ -324,8 +319,8 @@ func (pool *OrderPool) reset(oldHead, newblock *types.Block) {
 }
 
 // Stop terminates the transaction pool.
-func (pool *OrderPool) Stop() {
-	// Unsubscribe all subscriptions registered from OrderPool
+func (pool *LendingPool) Stop() {
+	// Unsubscribe all subscriptions registered from LendingPool
 	pool.scope.Close()
 
 	// Unsubscribe subscriptions registered from blockchain
@@ -340,12 +335,12 @@ func (pool *OrderPool) Stop() {
 
 // SubscribeTxPreEvent registers a subscription of TxPreEvent and
 // starts sending event to the given channel.
-func (pool *OrderPool) SubscribeTxPreEvent(ch chan<- OrderTxPreEvent) event.Subscription {
+func (pool *LendingPool) SubscribeTxPreEvent(ch chan<- core.LendingTxPreEvent) event.Subscription {
 	return pool.scope.Track(pool.txFeed.Subscribe(ch))
 }
 
 // State returns the virtual managed state of the transaction pool.
-func (pool *OrderPool) State() *tradingstate.XDCXManagedState {
+func (pool *LendingPool) State() *lendingstate.LendingManagedState {
 	pool.mu.RLock()
 	defer pool.mu.RUnlock()
 
@@ -354,7 +349,7 @@ func (pool *OrderPool) State() *tradingstate.XDCXManagedState {
 
 // Stats retrieves the current pool stats, namely the number of pending and the
 // number of queued (non-executable) transactions.
-func (pool *OrderPool) Stats() (int, int) {
+func (pool *LendingPool) Stats() (int, int) {
 	pool.mu.RLock()
 	defer pool.mu.RUnlock()
 
@@ -363,7 +358,7 @@ func (pool *OrderPool) Stats() (int, int) {
 
 // stats retrieves the current pool stats, namely the number of pending and the
 // number of queued (non-executable) transactions.
-func (pool *OrderPool) stats() (int, int) {
+func (pool *LendingPool) stats() (int, int) {
 	pending := 0
 	for _, list := range pool.pending {
 		pending += list.Len()
@@ -375,31 +370,14 @@ func (pool *OrderPool) stats() (int, int) {
 	return pending, queued
 }
 
-// Content retrieves the data content of the transaction pool, returning all the
-// pending as well as queued transactions, grouped by account and sorted by nonce.
-func (pool *OrderPool) Content() (map[common.Address]types.OrderTransactions, map[common.Address]types.OrderTransactions) {
-	pool.mu.Lock()
-	defer pool.mu.Unlock()
-
-	pending := make(map[common.Address]types.OrderTransactions)
-	for addr, list := range pool.pending {
-		pending[addr] = list.Flatten()
-	}
-	queued := make(map[common.Address]types.OrderTransactions)
-	for addr, list := range pool.queue {
-		queued[addr] = list.Flatten()
-	}
-	return pending, queued
-}
-
 // Pending retrieves all currently processable transactions, groupped by origin
 // account and sorted by nonce. The returned transaction set is a copy and can be
 // freely modified by calling code.
-func (pool *OrderPool) Pending() (map[common.Address]types.OrderTransactions, error) {
+func (pool *LendingPool) Pending() (map[common.Address]types.LendingTransactions, error) {
 	pool.mu.Lock()
 	defer pool.mu.Unlock()
 
-	pending := make(map[common.Address]types.OrderTransactions)
+	pending := make(map[common.Address]types.LendingTransactions)
 	for addr, list := range pool.pending {
 		pending[addr] = list.Flatten()
 	}
@@ -409,8 +387,8 @@ func (pool *OrderPool) Pending() (map[common.Address]types.OrderTransactions, er
 // local retrieves all currently known local transactions, groupped by origin
 // account and sorted by nonce. The returned transaction set is a copy and can be
 // freely modified by calling code.
-func (pool *OrderPool) local() map[common.Address]types.OrderTransactions {
-	txs := make(map[common.Address]types.OrderTransactions)
+func (pool *LendingPool) local() map[common.Address]types.LendingTransactions {
+	txs := make(map[common.Address]types.LendingTransactions)
 	for addr := range pool.locals.accounts {
 		if pending := pool.pending[addr]; pending != nil {
 			txs[addr] = append(txs[addr], pending.Flatten()...)
@@ -423,112 +401,226 @@ func (pool *OrderPool) local() map[common.Address]types.OrderTransactions {
 }
 
 // GetSender get sender from transaction
-func (pool *OrderPool) GetSender(tx *types.OrderTransaction) (common.Address, error) {
-	from, err := types.OrderSender(pool.signer, tx)
+func (pool *LendingPool) GetSender(tx *types.LendingTransaction) (common.Address, error) {
+	from, err := types.LendingSender(pool.signer, tx)
 	if err != nil {
 		return common.Address{}, ErrInvalidSender
 	}
 	return from, nil
 }
 
-func (pool *OrderPool) validateOrder(tx *types.OrderTransaction) error {
-	orderSide := tx.Side()
-	orderType := tx.Type()
-	orderStatus := tx.Status()
-	price := tx.Price()
+func (pool *LendingPool) validateNewLending(cloneStateDb *state.StateDB, cloneLendingStateDb *lendingstate.LendingStateDB, tx *types.LendingTransaction) error {
+	lendingSide := tx.Side()
+	lendingType := tx.Type()
+	interest := tx.Interest()
 	quantity := tx.Quantity()
-
-	cloneStateDb := pool.currentRootState.Copy()
-	cloneXDCXStateDb := pool.currentOrderState.Copy()
-
-	if !tx.IsCancelledOrder() {
-		if quantity == nil || quantity.Cmp(big.NewInt(0)) <= 0 {
-			return ErrInvalidOrderQuantity
+	if quantity == nil || quantity.Sign() <= 0 {
+		return ErrInvalidLendingQuantity
+	}
+	if lendingType != LendingTypeMarket {
+		if interest <= 0 {
+			return ErrInvalidLendingInterest
 		}
-		if orderType != OrderTypeMarket {
-			if price == nil || price.Cmp(big.NewInt(0)) <= 0 {
-				return ErrInvalidOrderPrice
+	}
+
+	if lendingSide != lendingstate.Investing && lendingSide != lendingstate.Borrowing {
+		return ErrInvalidLendingSide
+	}
+	if lendingType != LendingTypeLimit && lendingType != LendingTypeMarket {
+		return ErrInvalidLendingType
+	}
+	if tx.Side() == lendingstate.Borrowing {
+		if tx.CollateralToken().IsZero() || tx.CollateralToken() == tx.LendingToken() {
+			return ErrInvalidLendingCollateral
+		}
+		validCollateral := false
+		collateralList := lendingstate.GetCollaterals(cloneStateDb, tx.RelayerAddress(), tx.LendingToken(), tx.Term())
+		for _, collateral := range collateralList {
+			if tx.CollateralToken() == collateral {
+				validCollateral = true
+				break
 			}
 		}
-
-		if orderSide != OrderSideAsk && orderSide != OrderSideBid {
-			return ErrInvalidOrderSide
+		if !validCollateral {
+			return ErrInvalidLendingCollateral
 		}
-		if orderType != OrderTypeLimit && orderType != OrderTypeMarket {
-			return ErrInvalidOrderType
-		}
-		if err := tradingstate.VerifyPair(cloneStateDb, tx.ExchangeAddress(), tx.BaseToken(), tx.QuoteToken()); err != nil {
+	}
+	if lendingType == LendingTypeLimit {
+		if err := pool.validateBalance(cloneStateDb, cloneLendingStateDb, tx, tx.CollateralToken()); err != nil {
 			return err
 		}
+	}
+	return nil
+}
 
-		if orderType == OrderTypeLimit {
-			XDPoSEngine, ok := pool.chain.Engine().(*XDPoS.XDPoS)
-			if !ok {
-				return ErrNotXDPoS
-			}
-			XDCXServ := XDPoSEngine.GetXDCXService()
-			if XDCXServ == nil {
-				return errors.New("XDCx not found in order validation")
-			}
-			baseDecimal, err := XDCXServ.GetTokenDecimal(pool.chain, cloneStateDb, tx.BaseToken())
+func (pool *LendingPool) validateCancelledLending(cloneLendingStateDb *lendingstate.LendingStateDB, tx *types.LendingTransaction) error {
+	if tx.LendingId() == 0 {
+		return ErrInvalidCancelledLending
+	}
+	item := cloneLendingStateDb.GetLendingOrder(lendingstate.GetLendingOrderBookHash(tx.LendingToken(), tx.Term()), common.Uint64ToHash(tx.LendingId()))
+	if item == lendingstate.EmptyLendingOrder {
+		log.Debug("LendingOrder not found ", "LendingId", tx.LendingId(), "LendToken", tx.LendingToken().Hex(), "CollateralToken", tx.CollateralToken().Hex(), "Term", tx.Term())
+		return ErrInvalidCancelledLending
+	}
+	if item.Hash != tx.LendingHash() {
+		log.Debug("Invalid lending hash", "expected", item.Hash.Hex(), "got", tx.LendingHash().Hex())
+		return ErrInvalidLendingHash
+	}
+	return nil
+}
+func (pool *LendingPool) validateRepayLending(cloneStateDb *state.StateDB, cloneLendingStateDb *lendingstate.LendingStateDB, tx *types.LendingTransaction) error {
+	if tx.LendingTradeId() == 0 {
+		return ErrInvalidLendingTradeID
+	}
+	lendingBook := lendingstate.GetLendingOrderBookHash(tx.LendingToken(), tx.Term())
+	lendingTrade := cloneLendingStateDb.GetLendingTrade(lendingBook, common.Uint64ToHash(tx.LendingTradeId()))
+	if lendingTrade == lendingstate.EmptyLendingTrade {
+		return ErrInvalidLendingTradeID
+	}
+	if tx.UserAddress() != lendingTrade.Borrower {
+		return ErrInvalidLendingUserAddress
+	}
+	if tx.RelayerAddress() != lendingTrade.BorrowingRelayer {
+		return ErrInvalidLendingRelayer
+	}
+	if err := pool.validateBalance(cloneStateDb, cloneLendingStateDb, tx, tx.CollateralToken()); err != nil {
+		return err
+	}
+	return nil
+}
+func (pool *LendingPool) validateTopupLending(cloneStateDb *state.StateDB, cloneLendingStateDb *lendingstate.LendingStateDB, tx *types.LendingTransaction) error {
+	if tx.LendingTradeId() == 0 {
+		return ErrInvalidLendingTradeID
+	}
+	if tx.Quantity() == nil || tx.Quantity().Sign() <= 0 {
+		return ErrInvalidLendingQuantity
+	}
+	lendingBook := lendingstate.GetLendingOrderBookHash(tx.LendingToken(), tx.Term())
+	lendingTrade := cloneLendingStateDb.GetLendingTrade(lendingBook, common.Uint64ToHash(tx.LendingTradeId()))
+	if lendingTrade == lendingstate.EmptyLendingTrade {
+		return ErrInvalidLendingTradeID
+	}
+	if tx.UserAddress() != lendingTrade.Borrower {
+		return ErrInvalidLendingUserAddress
+	}
+	if tx.RelayerAddress() != lendingTrade.BorrowingRelayer {
+		return ErrInvalidLendingRelayer
+	}
+	if err := pool.validateBalance(cloneStateDb, cloneLendingStateDb, tx, lendingTrade.CollateralToken); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (pool *LendingPool) validateBalance(cloneStateDb *state.StateDB, cloneLendingStateDb *lendingstate.LendingStateDB, tx *types.LendingTransaction, collateralToken common.Address) error {
+	XDPoSEngine, ok := pool.chain.Engine().(*XDPoS.XDPoS)
+	if !ok {
+		return core.ErrNotXDPoS
+	}
+	XDCXServ := XDPoSEngine.GetXDCXService()
+	lendingServ := XDPoSEngine.GetLendingService()
+	if XDCXServ == nil {
+		return errors.New("XDCx not found in order validation")
+	}
+	lendingTokenDecimal, err := XDCXServ.GetTokenDecimal(pool.chain, cloneStateDb, tx.LendingToken())
+	if err != nil {
+		return fmt.Errorf("validateOrder: failed to get lendingTokenDecimal. err: %v", err)
+	}
+	author, err := pool.chain.Engine().Author(pool.chain.CurrentHeader())
+	if err != nil {
+		return err
+	}
+	tradingStateDb, err := XDCXServ.GetTradingState(pool.chain.CurrentBlock(), author)
+	if err != nil {
+		return fmt.Errorf("validateLending: failed to get tradingStateDb. Error: %v", err)
+	}
+	cloneTradingStateDb := tradingStateDb.Copy()
+	// collateralPrice: price of collateral by LendingToken
+	// Eg: LendingToken: USD, CollateralToken: BTC
+	// collateralPrice = BTC/USD (eg: 8000 USD)
+	// lendTokenXDCPrice: price of lendingToken in XDC quote
+	var lendTokenXDCPrice, collateralPrice, collateralTokenDecimal *big.Int
+	if !collateralToken.IsZero() {
+		collateralTokenDecimal, err = XDCXServ.GetTokenDecimal(pool.chain, cloneStateDb, collateralToken)
+		if err != nil {
+			return fmt.Errorf("validateOrder: failed to get collateralTokenDecimal. err: %v", err)
+		}
+		lendTokenXDCPrice, collateralPrice, err = lendingServ.GetCollateralPrices(pool.chain.CurrentHeader(), pool.chain, cloneStateDb, cloneTradingStateDb, collateralToken, tx.LendingToken())
+		if err != nil {
+			return err
+		}
+		if lendTokenXDCPrice == nil || lendTokenXDCPrice.Sign() <= 0 || collateralPrice == nil || collateralPrice.Sign() <= 0 {
+			log.Debug("ValidateLending: ErrInvalidCollateralPrice", "lendTokenXDCPrice", lendTokenXDCPrice, "collateralPrice", collateralPrice)
+			return lendingstate.ErrInvalidCollateralPrice
+		}
+	}
+	if lendTokenXDCPrice == nil || lendTokenXDCPrice.Sign() == 0 {
+		if tx.LendingToken() == common.XDCNativeAddressBinary {
+			lendTokenXDCPrice = common.BasePrice
+		} else {
+			lendTokenXDCPrice, err = lendingServ.GetMediumTradePriceBeforeEpoch(pool.chain, cloneStateDb, cloneTradingStateDb, tx.LendingToken(), common.XDCNativeAddressBinary)
 			if err != nil {
-				return fmt.Errorf("validateOrder: failed to get baseDecimal. err: %v", err)
-			}
-			quoteDecimal, err := XDCXServ.GetTokenDecimal(pool.chain, cloneStateDb, tx.QuoteToken())
-			if err != nil {
-				return fmt.Errorf("validateOrder: failed to get quoteDecimal. err: %v", err)
-			}
-			if err := tradingstate.VerifyBalance(cloneStateDb, cloneXDCXStateDb, tx, baseDecimal, quoteDecimal); err != nil {
 				return err
 			}
 		}
-
 	}
-
-	if orderStatus != OrderStatusNew && orderStatus != OrderStatusCancle {
-		return ErrInvalidOrderStatus
+	isXDCXLendingFork := pool.chain.Config().IsTIPXDCXLending(pool.chain.CurrentHeader().Number)
+	if err := lendingstate.VerifyBalance(isXDCXLendingFork,
+		cloneStateDb,
+		cloneLendingStateDb,
+		tx.Type(),
+		tx.Side(),
+		tx.Status(),
+		tx.UserAddress(),
+		tx.RelayerAddress(),
+		tx.LendingToken(),
+		tx.CollateralToken(),
+		tx.Quantity(),
+		lendingTokenDecimal,
+		collateralTokenDecimal,
+		lendTokenXDCPrice,
+		collateralPrice,
+		tx.Term(),
+		tx.LendingId(),
+		tx.LendingTradeId(),
+	); err != nil {
+		return err
 	}
-	var signer = types.OrderTxSigner{}
-
-	if !tx.IsCancelledOrder() {
-		if !common.EmptyHash(tx.OrderHash()) {
-			if signer.Hash(tx) != tx.OrderHash() {
-				return ErrInvalidOrderHash
-			}
-		} else {
-			tx.SetOrderHash(signer.Hash(tx))
-		}
-
-	} else {
-		if tx.OrderID() == 0 {
-			return ErrInvalidCancelledOrder
-		}
-		originOrder := cloneXDCXStateDb.GetOrder(tradingstate.GetTradingOrderBookHash(tx.BaseToken(), tx.QuoteToken()), common.BigToHash(new(big.Int).SetUint64(tx.OrderID())))
-		if originOrder == tradingstate.EmptyOrder {
-			log.Debug("Order not found ", "OrderId", tx.OrderID(), "BaseToken", tx.BaseToken().Hex(), "QuoteToken", tx.QuoteToken().Hex())
-			return ErrInvalidCancelledOrder
-		}
-		if originOrder.Hash != tx.OrderHash() {
-			log.Debug("Invalid order hash", "expected", originOrder.Hash.Hex(), "got", tx.OrderHash().Hex())
-			return ErrInvalidOrderHash
-		}
-	}
-
-	from, _ := types.OrderSender(pool.signer, tx)
-	if from != tx.UserAddress() {
-		return ErrInvalidOrderUserAddress
-	}
-
-	if !tradingstate.IsValidRelayer(cloneStateDb, tx.ExchangeAddress()) {
-		return fmt.Errorf("invalid relayer. ExchangeAddress: %s", tx.ExchangeAddress().Hex())
-	}
-
 	return nil
+}
+
+func (pool *LendingPool) validateLending(tx *types.LendingTransaction) error {
+	cloneStateDb := pool.currentRootState.Copy()
+	cloneLendingStateDb := pool.currentLendingState.Copy()
+	from, _ := types.LendingSender(pool.signer, tx)
+	if from != tx.UserAddress() {
+		return ErrInvalidLendingUserAddress
+	}
+	if !lendingstate.IsValidRelayer(cloneStateDb, tx.RelayerAddress()) {
+		return fmt.Errorf("invalid lending relayer. ExchangeAddress: %s", tx.RelayerAddress().Hex())
+	}
+	if valid, _ := lendingstate.IsValidPair(cloneStateDb, tx.RelayerAddress(), tx.LendingToken(), tx.Term()); !valid {
+		return fmt.Errorf("invalid pair. Relayer: %s. LendingToken: %s. Term: %d", tx.RelayerAddress().Hex(), tx.LendingToken().Hex(), tx.Term())
+	}
+	if tx.IsCreatedLending() {
+		return pool.validateNewLending(cloneStateDb, cloneLendingStateDb, tx)
+	}
+	if tx.IsCancelledLending() {
+		return pool.validateCancelledLending(cloneLendingStateDb, tx)
+	}
+	if tx.IsTopupLending() {
+		return pool.validateTopupLending(cloneStateDb, cloneLendingStateDb, tx)
+	}
+	if tx.IsRepayLending() {
+		return pool.validateRepayLending(cloneStateDb, cloneLendingStateDb, tx)
+	}
+
+	return ErrInvalidLendingStatus
 }
 
 // validateTx checks whether a transaction is valid according to the consensus
 // rules and adheres to some heuristic limits of the local node (price and size).
-func (pool *OrderPool) validateTx(tx *types.OrderTransaction, local bool) error {
+func (pool *LendingPool) validateTx(tx *types.LendingTransaction, local bool) error {
 
 	// check if sender is in black list
 	if tx.From() != nil && common.Blacklist[*tx.From()] {
@@ -540,20 +632,20 @@ func (pool *OrderPool) validateTx(tx *types.OrderTransaction, local bool) error 
 	}
 
 	// Make sure the transaction is signed properly
-	from, err := types.OrderSender(pool.signer, tx)
+	from, err := types.LendingSender(pool.signer, tx)
 	if err != nil {
 		return ErrInvalidSender
 	}
-	err = pool.validateOrder(tx)
+	err = pool.validateLending(tx)
 	if err != nil {
 		return err
 	}
-	// Ensure the transaction adheres to nonce ordering
-	if pool.currentOrderState.GetNonce(from.Hash()) > tx.Nonce() {
-		return ErrNonceTooLow
+	// Ensure the transaction adheres to nonce lending
+	if pool.currentLendingState.GetNonce(from.Hash()) > tx.Nonce() {
+		return core.ErrNonceTooLow
 	}
 	if pool.pendingState.GetNonce(from.Hash())+common.LimitThresholdNonceInQueue < tx.Nonce() {
-		return ErrNonceTooHigh
+		return core.ErrNonceTooHigh
 	}
 
 	return nil
@@ -567,25 +659,25 @@ func (pool *OrderPool) validateTx(tx *types.OrderTransaction, local bool) error 
 // If a newly added transaction is marked as local, its sending account will be
 // whitelisted, preventing any associated transaction from being dropped out of
 // the pool due to pricing constraints.
-func (pool *OrderPool) add(tx *types.OrderTransaction, local bool) (bool, error) {
+func (pool *LendingPool) add(tx *types.LendingTransaction, local bool) (bool, error) {
 	// If the transaction is already known, discard it
 	hash := tx.Hash()
 	if pool.all[hash] != nil {
-		log.Debug("Discarding known order transaction", "hash", hash, "userAddress", tx.UserAddress().Hex(), "status", tx.Status)
+		log.Debug("Discarding already known transaction", "hash", hash)
 		return false, fmt.Errorf("known transaction: %x", hash)
 	}
 
 	// If the transaction fails basic validation, discard it
 	if err := pool.validateTx(tx, local); err != nil {
-		log.Debug("Discarding invalid order transaction", "hash", hash, "userAddress", tx.UserAddress().Hex(), "status", tx.Status, "err", err)
+		log.Debug("Discarding invalid lending transaction", "hash", hash, "userAddress", tx.UserAddress, "status", tx.Status, "err", err)
 		invalidTxMeter.Mark(1)
 		return false, err
 	}
-	from, _ := types.OrderSender(pool.signer, tx) // already validated
+	from, _ := types.LendingSender(pool.signer, tx) // already validated
 
 	// If the transaction pool is full, discard underpriced transactions
 	if uint64(len(pool.all)) >= pool.config.GlobalSlots+pool.config.GlobalQueue {
-		log.Debug("Add order transaction to pool full", "hash", hash, "nonce", tx.Nonce())
+		log.Debug("Add lending transaction to pool full", "hash", hash, "nonce", tx.Nonce())
 		return false, ErrPoolOverflow
 	}
 	// If the transaction is replacing an already pending one, do directly
@@ -602,8 +694,7 @@ func (pool *OrderPool) add(tx *types.OrderTransaction, local bool) (bool, error)
 		pool.all[tx.Hash()] = tx
 		pool.journalTx(from, tx)
 
-		log.Debug("Pooled new executable transaction", "hash", hash, "useraddress", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "status", tx.Status(), "orderid", tx.OrderID())
-		go pool.txFeed.Send(OrderTxPreEvent{tx})
+		log.Debug("Lending Pooled new executable transaction", "hash", hash, "useraddress", tx.UserAddress(), "nonce", tx.Nonce(), "status", tx.Status(), "lendingid", tx.LendingId())
 		return old != nil, nil
 
 	}
@@ -625,17 +716,16 @@ func (pool *OrderPool) add(tx *types.OrderTransaction, local bool) (bool, error)
 // enqueueTx inserts a new transaction into the non-executable transaction queue.
 //
 // Note, this method assumes the pool lock is held!
-func (pool *OrderPool) enqueueTx(hash common.Hash, tx *types.OrderTransaction) (bool, error) {
+func (pool *LendingPool) enqueueTx(hash common.Hash, tx *types.LendingTransaction) (bool, error) {
 	// Try to insert the transaction into the future queue
-	log.Debug("enqueueTx", "hash", hash, "useraddress", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "status", tx.Status(), "orderid", tx.OrderID())
-	from, _ := types.OrderSender(pool.signer, tx) // already validated
+	from, _ := types.LendingSender(pool.signer, tx) // already validated
 	if pool.queue[from] == nil {
-		pool.queue[from] = newOrderTxList(false)
+		pool.queue[from] = newLendingTxList(false)
 	}
 	inserted, old := pool.queue[from].Add(tx)
 	if !inserted {
 		// An older transaction was better, discard this
-		queuedDiscardMeter.Mark(1)
+		pendingDiscardMeter.Mark(1)
 		return false, ErrPendingNonceTooLow
 	}
 	// Discard any previous transaction and mark this
@@ -649,7 +739,7 @@ func (pool *OrderPool) enqueueTx(hash common.Hash, tx *types.OrderTransaction) (
 
 // journalTx adds the specified transaction to the local disk journal if it is
 // deemed to have been sent from a local account.
-func (pool *OrderPool) journalTx(from common.Address, tx *types.OrderTransaction) {
+func (pool *LendingPool) journalTx(from common.Address, tx *types.LendingTransaction) {
 	// Only journal if it's enabled and the transaction is local
 	if pool.journal == nil || !pool.locals.contains(from) {
 		return
@@ -662,11 +752,10 @@ func (pool *OrderPool) journalTx(from common.Address, tx *types.OrderTransaction
 // promoteTx adds a transaction to the pending (processable) list of transactions.
 //
 // Note, this method assumes the pool lock is held!
-func (pool *OrderPool) promoteTx(addr common.Address, hash common.Hash, tx *types.OrderTransaction) {
+func (pool *LendingPool) promoteTx(addr common.Address, hash common.Hash, tx *types.LendingTransaction) {
 	// Try to insert the transaction into the pending queue
-	log.Debug("promoteTx", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
 	if pool.pending[addr] == nil {
-		pool.pending[addr] = newOrderTxList(true)
+		pool.pending[addr] = newLendingTxList(true)
 	}
 	list := pool.pending[addr]
 
@@ -689,50 +778,46 @@ func (pool *OrderPool) promoteTx(addr common.Address, hash common.Hash, tx *type
 	// Set the potentially new pending nonce and notify any subsystems of the new tx
 	pool.beats[addr] = time.Now()
 	pool.pendingState.SetNonce(addr.Hash(), tx.Nonce()+1)
-	log.Debug("promoteTx txFeed.Send", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
-	go pool.txFeed.Send(OrderTxPreEvent{tx})
+
+	go pool.txFeed.Send(core.LendingTxPreEvent{Tx: tx})
 }
 
 // AddLocal enqueues a single transaction into the pool if it is valid, marking
 // the sender as a local one in the mean time, ensuring it goes around the local
 // pricing constraints.
-func (pool *OrderPool) AddLocal(tx *types.OrderTransaction) error {
-	log.Debug("AddLocal order add local tx", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+func (pool *LendingPool) AddLocal(tx *types.LendingTransaction) error {
+	log.Debug("Lending add local tx", "relayeraddress", tx.RelayerAddress().Hex(), "addr", tx.UserAddress(), "nonce", tx.Nonce(), "ohash", tx.LendingHash().Hex(), "status", tx.Status(), "lendingid", tx.LendingId(), "lendingtradeid", tx.LendingTradeId())
 	return pool.addTx(tx, !pool.config.NoLocals)
 }
 
 // AddRemote enqueues a single transaction into the pool if it is valid. If the
 // sender is not among the locally tracked ones, full pricing constraints will
 // apply.
-func (pool *OrderPool) AddRemote(tx *types.OrderTransaction) error {
-	log.Debug("AddRemote", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+func (pool *LendingPool) AddRemote(tx *types.LendingTransaction) error {
 	return pool.addTx(tx, false)
 }
 
 // AddLocals enqueues a batch of transactions into the pool if they are valid,
 // marking the senders as a local ones in the mean time, ensuring they go around
 // the local pricing constraints.
-func (pool *OrderPool) AddLocals(txs []*types.OrderTransaction) []error {
+func (pool *LendingPool) AddLocals(txs []*types.LendingTransaction) []error {
 	return pool.addTxs(txs, !pool.config.NoLocals)
 }
 
 // AddRemotes enqueues a batch of transactions into the pool if they are valid.
 // If the senders are not among the locally tracked ones, full pricing constraints
 // will apply.
-func (pool *OrderPool) AddRemotes(txs []*types.OrderTransaction) []error {
-	for _, tx := range txs {
-		log.Debug("AddRemotes", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
-	}
+func (pool *LendingPool) AddRemotes(txs []*types.LendingTransaction) []error {
 	return pool.addTxs(txs, false)
 }
 
 // addTx enqueues a single transaction into the pool if it is valid.
-func (pool *OrderPool) addTx(tx *types.OrderTransaction, local bool) error {
+func (pool *LendingPool) addTx(tx *types.LendingTransaction, local bool) error {
 	if !pool.chainconfig.IsTIPXDCXReceiver(pool.chain.CurrentBlock().Number()) {
 		return nil
 	}
 	tx.CacheHash()
-	types.CacheOrderSigner(pool.signer, tx)
+	types.CacheLendingSigner(pool.signer, tx)
 	pool.mu.Lock()
 	defer pool.mu.Unlock()
 
@@ -743,14 +828,14 @@ func (pool *OrderPool) addTx(tx *types.OrderTransaction, local bool) error {
 	}
 	// If we added a new transaction, run promotion checks and return
 	if !replace {
-		from, _ := types.OrderSender(pool.signer, tx) // already validated
+		from, _ := types.LendingSender(pool.signer, tx) // already validated
 		pool.promoteExecutables([]common.Address{from})
 	}
 	return nil
 }
 
 // addTxs attempts to queue a batch of transactions if they are valid.
-func (pool *OrderPool) addTxs(txs []*types.OrderTransaction, local bool) []error {
+func (pool *LendingPool) addTxs(txs []*types.LendingTransaction, local bool) []error {
 	pool.mu.Lock()
 	defer pool.mu.Unlock()
 
@@ -759,7 +844,7 @@ func (pool *OrderPool) addTxs(txs []*types.OrderTransaction, local bool) []error
 
 // addTxsLocked attempts to queue a batch of transactions if they are valid,
 // whilst assuming the transaction pool lock is already held.
-func (pool *OrderPool) addTxsLocked(txs []*types.OrderTransaction, local bool) []error {
+func (pool *LendingPool) addTxsLocked(txs []*types.LendingTransaction, local bool) []error {
 	// Add the batch of transaction, tracking the accepted ones
 	dirty := make(map[common.Address]struct{})
 	errs := make([]error, len(txs))
@@ -768,7 +853,7 @@ func (pool *OrderPool) addTxsLocked(txs []*types.OrderTransaction, local bool) [
 		var replace bool
 		if replace, errs[i] = pool.add(tx, local); errs[i] == nil {
 			if !replace {
-				from, _ := types.OrderSender(pool.signer, tx) // already validated
+				from, _ := types.LendingSender(pool.signer, tx) // already validated
 				dirty[from] = struct{}{}
 			}
 		}
@@ -786,14 +871,14 @@ func (pool *OrderPool) addTxsLocked(txs []*types.OrderTransaction, local bool) [
 
 // Status returns the status (unknown/pending/queued) of a batch of transactions
 // identified by their hashes.
-func (pool *OrderPool) Status(hashes []common.Hash) []TxStatus {
+func (pool *LendingPool) Status(hashes []common.Hash) []TxStatus {
 	pool.mu.RLock()
 	defer pool.mu.RUnlock()
 
 	status := make([]TxStatus, len(hashes))
 	for i, hash := range hashes {
 		if tx := pool.all[hash]; tx != nil {
-			from, _ := types.OrderSender(pool.signer, tx) // already validated
+			from, _ := types.LendingSender(pool.signer, tx) // already validated
 			if pool.pending[from] != nil && pool.pending[from].txs.items[tx.Nonce()] != nil {
 				status[i] = TxStatusPending
 			} else {
@@ -806,7 +891,7 @@ func (pool *OrderPool) Status(hashes []common.Hash) []TxStatus {
 
 // Get returns a transaction if it is contained in the pool
 // and nil otherwise.
-func (pool *OrderPool) Get(hash common.Hash) *types.OrderTransaction {
+func (pool *LendingPool) Get(hash common.Hash) *types.LendingTransaction {
 	pool.mu.RLock()
 	defer pool.mu.RUnlock()
 
@@ -815,13 +900,13 @@ func (pool *OrderPool) Get(hash common.Hash) *types.OrderTransaction {
 
 // removeTx removes a single transaction from the queue, moving all subsequent
 // transactions back to the future queue.
-func (pool *OrderPool) removeTx(hash common.Hash) {
+func (pool *LendingPool) removeTx(hash common.Hash) {
 	// Fetch the transaction we wish to delete
 	tx, ok := pool.all[hash]
 	if !ok {
 		return
 	}
-	addr, _ := types.OrderSender(pool.signer, tx) // already validated during insertion
+	addr, _ := types.LendingSender(pool.signer, tx) // already validated during insertion
 
 	// Remove it from the list of known transactions
 	delete(pool.all, hash)
@@ -857,11 +942,11 @@ func (pool *OrderPool) removeTx(hash common.Hash) {
 // promoteExecutables moves transactions that have become processable from the
 // future queue to the set of pending transactions. During this process, all
 // invalidated transactions (low nonce, low balance) are deleted.
-func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
+func (pool *LendingPool) promoteExecutables(accounts []common.Address) {
+	log.Debug("start promoteExecutables")
 	defer func(start time.Time) {
 		log.Debug("end promoteExecutables", "time", common.PrettyDuration(time.Since(start)))
 	}(time.Now())
-
 	// Gather all the accounts potentially needing updates
 	if accounts == nil {
 		accounts = make([]common.Address, 0, len(pool.queue))
@@ -876,9 +961,9 @@ func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
 			continue // Just in case someone calls with a non existing account
 		}
 		// Drop all transactions that are deemed too old (low nonce)
-		for _, tx := range list.Forward(pool.currentOrderState.GetNonce(addr.Hash())) {
+		for _, tx := range list.Forward(pool.currentLendingState.GetNonce(addr.Hash())) {
 			hash := tx.Hash()
-			log.Debug("Removed old queued transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+			log.Trace("Removed old queued transaction", "hash", hash)
 			delete(pool.all, hash)
 
 		}
@@ -886,7 +971,7 @@ func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
 		// Gather all executable transactions and promote them
 		for _, tx := range list.Ready(pool.pendingState.GetNonce(addr.Hash())) {
 			hash := tx.Hash()
-			log.Debug("Promoting queued transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+			log.Trace("Promoting queued transaction", "hash", hash)
 			pool.promoteTx(addr, hash, tx)
 		}
 		// Drop all transactions over the allowed limit
@@ -896,12 +981,11 @@ func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
 				delete(pool.all, hash)
 
 				queuedRateLimitMeter.Mark(1)
-				log.Debug("Removed cap-exceeding queued transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+				log.Trace("Removed cap-exceeding queued transaction", "hash", hash)
 			}
 		}
 		// Delete the entire queue entry if it became empty.
 		if list.Empty() {
-			log.Debug("promoteExecutables remove transaction queue", "addr", addr.Hex())
 			delete(pool.queue, addr)
 		}
 	}
@@ -945,7 +1029,7 @@ func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
 							if nonce := tx.Nonce(); pool.pendingState.GetNonce(offenders[i].Hash()) > nonce {
 								pool.pendingState.SetNonce(offenders[i].Hash(), nonce)
 							}
-							log.Debug("Removed fairness-exceeding pending transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+							log.Trace("Removed fairness-exceeding pending transaction", "hash", hash)
 						}
 						pending--
 					}
@@ -966,7 +1050,7 @@ func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
 						if nonce := tx.Nonce(); pool.pendingState.GetNonce(addr.Hash()) > nonce {
 							pool.pendingState.SetNonce(addr.Hash(), nonce)
 						}
-						log.Debug("Removed fairness-exceeding pending transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+						log.Trace("Removed fairness-exceeding pending transaction", "hash", hash)
 					}
 					pending--
 				}
@@ -1019,15 +1103,15 @@ func (pool *OrderPool) promoteExecutables(accounts []common.Address) {
 // demoteUnexecutables removes invalid and processed transactions from the pools
 // executable/pending queue and any subsequent transactions that become unexecutable
 // are moved back into the future queue.
-func (pool *OrderPool) demoteUnexecutables() {
+func (pool *LendingPool) demoteUnexecutables() {
 	// Iterate over all accounts and demote any non-executable transactions
 	for addr, list := range pool.pending {
-		nonce := pool.currentOrderState.GetNonce(addr.Hash())
+		nonce := pool.currentLendingState.GetNonce(addr.Hash())
 		log.Debug("demoteUnexecutables", "addr", addr.Hex(), "nonce", nonce)
 		// Drop all transactions that are deemed too old (low nonce)
 		for _, tx := range list.Forward(nonce) {
 			hash := tx.Hash()
-			log.Debug("demoteUnexecutables removed old queued transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+			log.Debug("Removed old pending transaction", "hash", hash)
 			delete(pool.all, hash)
 		}
 
@@ -1035,7 +1119,7 @@ func (pool *OrderPool) demoteUnexecutables() {
 		if list.Len() > 0 && list.txs.Get(nonce) == nil {
 			for _, tx := range list.Cap(0) {
 				hash := tx.Hash()
-				log.Debug("demoteUnexecutables Demoting invalidated transaction", "addr", tx.UserAddress().Hex(), "nonce", tx.Nonce(), "ohash", tx.OrderHash().Hex(), "status", tx.Status(), "orderid", tx.OrderID())
+				log.Warn("Demoting invalidated transaction", "hash", hash)
 				pool.enqueueTx(hash, tx)
 			}
 		}
@@ -1047,36 +1131,36 @@ func (pool *OrderPool) demoteUnexecutables() {
 	}
 }
 
-type orderAccountSet struct {
+type lendingAccountSet struct {
 	accounts map[common.Address]struct{}
-	signer   types.OrderSigner
+	signer   types.LendingSigner
 }
 
 // newAccountSet creates a new address set with an associated signer for sender
 // derivations.
-func newOrderAccountSet(signer types.OrderSigner) *orderAccountSet {
-	return &orderAccountSet{
+func newLendingAccountSet(signer types.LendingSigner) *lendingAccountSet {
+	return &lendingAccountSet{
 		accounts: make(map[common.Address]struct{}),
 		signer:   signer,
 	}
 }
 
 // contains checks if a given address is contained within the set.
-func (as *orderAccountSet) contains(addr common.Address) bool {
+func (as *lendingAccountSet) contains(addr common.Address) bool {
 	_, exist := as.accounts[addr]
 	return exist
 }
 
 // containsTx checks if the sender of a given tx is within the set. If the sender
 // cannot be derived, this method returns false.
-func (as *orderAccountSet) containsTx(tx *types.OrderTransaction) bool {
-	if addr, err := types.OrderSender(as.signer, tx); err == nil {
+func (as *lendingAccountSet) containsTx(tx *types.LendingTransaction) bool {
+	if addr, err := types.LendingSender(as.signer, tx); err == nil {
 		return as.contains(addr)
 	}
 	return false
 }
 
 // add inserts a new address into the set to track.
-func (as *orderAccountSet) add(addr common.Address) {
+func (as *lendingAccountSet) add(addr common.Address) {
 	as.accounts[addr] = struct{}{}
 }
