+++ core/txpool/lending_pool_test.go
@@ -1,4 +1,4 @@
-package core
+package txpool
 
 import (
 	"context"
