+++ eth/ethconfig/gen_config.go
@@ -9,6 +9,7 @@ import (
 	"github.com/XinFinOrg/XDPoSChain/common"
 	"github.com/XinFinOrg/XDPoSChain/consensus/ethash"
 	"github.com/XinFinOrg/XDPoSChain/core"
+	"github.com/XinFinOrg/XDPoSChain/core/txpool"
 	"github.com/XinFinOrg/XDPoSChain/eth/downloader"
 	"github.com/XinFinOrg/XDPoSChain/eth/gasprice"
 )
@@ -33,7 +34,7 @@ func (c Config) MarshalTOML() (interface{}, error) {
 		GasPrice                *big.Int
 		FilterLogCacheSize      int
 		Ethash                  ethash.Config
-		TxPool                  core.TxPoolConfig
+		TxPool                  txpool.Config
 		GPO                     gasprice.Config
 		EnablePreimageRecording bool
 		DocRoot                 string `toml:"-"`
@@ -87,7 +88,7 @@ func (c *Config) UnmarshalTOML(unmarshal func(interface{}) error) error {
 		GasPrice                *big.Int
 		FilterLogCacheSize      *int
 		Ethash                  *ethash.Config
-		TxPool                  *core.TxPoolConfig
+		TxPool                  *txpool.Config
 		GPO                     *gasprice.Config
 		EnablePreimageRecording *bool
 		DocRoot                 *string `toml:"-"`
