+++ les/protocol.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/XinFinOrg/XDPoSChain/common"
 	"github.com/XinFinOrg/XDPoSChain/core"
+	"github.com/XinFinOrg/XDPoSChain/core/txpool"
 	"github.com/XinFinOrg/XDPoSChain/crypto"
 	"github.com/XinFinOrg/XDPoSChain/crypto/secp256k1"
 	"github.com/XinFinOrg/XDPoSChain/rlp"
@@ -223,7 +224,7 @@ type CodeData []struct {
 type proofsData [][]rlp.RawValue
 
 type txStatus struct {
-	Status core.TxStatus
+	Status txpool.TxStatus
 	Lookup *core.TxLookupEntry `rlp:"nil"`
 	Error  string
 }
